+++ src/app/api-server/src/config.rs
@@ -207,7 +207,7 @@ pub struct RepoCachingConfig {
 // UrlConfig
 ////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 
-#[derive(Debug, Serialize, Deserialize)]
+#[derive(Debug, Clone, Serialize, Deserialize)]
 #[serde(deny_unknown_fields)]
 #[serde(rename_all = "camelCase")]
 pub struct UrlConfig {