+++ CHANGELOG.md
@@ -13,9 +13,14 @@ Recommendation: for ease of reading, use the following order:
 -->
 
 ## [Unreleased]
+### Added
+- E2E: running tests also for S3 repositories
 ### Changed
 - Replaced default GraphQL playground with better maintained `graphiql` (old playground is still available)
 - Improved API server web console looks
+### Fixed
+- Flow progress notifier is now more resilient to deleted datasets
+- Use actual `base_url` in catalog configuration instead default one
 
 ## [0.56.1] - 2025-03-14
 ### Fixed